apps/streamlit_lakebase/requirements.txt
streamlit==1.39.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
databricks-sdk==0.33.0
